Feature: Add device trigger support (#25)
Adds support for device triggers

Currently tested with two different variants of `WPH-01`, a double side
and a single sided variant. The change introduces a different concept of
creating an MQTT device, where it returns a list of entities instead of
one entity, as some devices can have multiple entities.